% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{.build_locale_lookup_tables}
\alias{.build_locale_lookup_tables}
\title{Build per-locale lookup environments from a FUNCTIONS data frame.}
\usage{
.build_locale_lookup_tables(df)
}
\description{
Build per-locale lookup environments from a FUNCTIONS data frame.
}
\keyword{internal}
