% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlex.R
\name{.collect_labels}
\alias{.collect_labels}
\title{Collect labels in the same DFS order as .render_tree produces lines.}
\usage{
.collect_labels(nodes)
}
\description{
Collect labels in the same DFS order as .render_tree produces lines.
}
\keyword{internal}
