% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{.lambda_scope_close}
\alias{.lambda_scope_close}
\title{Decrease paren depth; pop the scope when depth reaches 0.}
\usage{
.lambda_scope_close(scope)
}
\description{
Decrease paren depth; pop the scope when depth reaches 0.
}
\keyword{internal}
