% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{.lambda_scope_open}
\alias{.lambda_scope_open}
\title{Increase paren depth of the innermost LAMBDA/LET scope.}
\usage{
.lambda_scope_open(scope)
}
\description{
Increase paren depth of the innermost LAMBDA/LET scope.
}
\keyword{internal}
