% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{.reset_terminology_state}
\alias{.reset_terminology_state}
\title{Reset locale state to "no terminology".}
\usage{
.reset_terminology_state()
}
\description{
Reset locale state to "no terminology".
}
\keyword{internal}
