% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{.stop_no_terminology}
\alias{.stop_no_terminology}
\title{Stop with a clear message when locale work is requested but no cache.}
\usage{
.stop_no_terminology()
}
\description{
Stop with a clear message when locale work is requested but no cache.
}
\keyword{internal}
