% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexer.R
\name{.tokenise}
\alias{.tokenise}
\title{Tokenise a formula string}
\usage{
.tokenise(formula, sep = ",")
}
\arguments{
\item{formula}{Character scalar, optionally starting with '='.}

\item{sep}{The argument separator (e.g., ',' or ';').}
}
\value{
A list of token objects, each with fields \code{type} and \code{val}.
}
\description{
Tokenise a formula string
}
\keyword{internal}
