% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{from_xml}
\alias{from_xml}
\title{Convert a OOXML formula to a user-facing Excel formula}
\usage{
from_xml(formula, locale = NULL)
}
\arguments{
\item{formula}{a formula}

\item{locale}{a locale}
}
\description{
Convert a OOXML formula to a user-facing Excel formula
}
