% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{from_xml_v}
\alias{from_xml_v}
\title{Translate a vector of OOXML formulas to user-facing Excel format}
\usage{
from_xml_v(formulas, locale = NULL)
}
\arguments{
\item{formulas}{Character vector of OOXML formulas.}

\item{locale}{Target locale code, or NULL.}
}
\value{
Character vector of Excel formulas.
}
\description{
Vectorised wrapper around \code{\link[=from_xml]{from_xml()}}.
}
\examples{
from_xml_v(c("=_xlfn.SEQUENCE(5)", "=SUM(_xlfn.ANCHORARRAY(A1))"))
}
