% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{has_terminology}
\alias{has_terminology}
\title{Check whether locale terminology has been loaded.}
\usage{
has_terminology()
}
\value{
Logical scalar.
}
\description{
Check whether locale terminology has been loaded.
}
\examples{
has_terminology()
}
\seealso{
\code{\link[=setup_terminology]{setup_terminology()}}, \code{\link[=clear_terminology]{clear_terminology()}}
}
