% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locales.R
\name{supported_locales}
\alias{supported_locales}
\title{List supported locales}
\usage{
supported_locales()
}
\value{
Character vector of supported locale codes.
}
\description{
List supported locales
}
\examples{
supported_locales()
}
