% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{to_xml}
\alias{to_xml}
\title{Convert a user-facing Excel formula to OOXML storage format}
\usage{
to_xml(formula, locale = NULL, warn_unknown = TRUE)
}
\arguments{
\item{formula}{a formula}

\item{locale}{a locale}

\item{warn_unknown}{a logical}
}
\description{
Convert a user-facing Excel formula to OOXML storage format
}
