% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_xml_v}
\alias{to_xml_v}
\title{Translate a vector of Excel formulas to OOXML storage format}
\usage{
to_xml_v(formulas, locale = NULL, warn_unknown = TRUE)
}
\arguments{
\item{formulas}{Character vector of Excel formulas.}

\item{locale}{Locale code, or NULL. Applied to all formulas.}

\item{warn_unknown}{Logical.}
}
\value{
Character vector of OOXML formulas.
}
\description{
Vectorised wrapper around \code{\link[=to_xml]{to_xml()}}.
}
\examples{
to_xml_v(c("=SUM(A1:A10)", "=SEQUENCE(5)", "=FILTER(A1:A10, B1:B10 > 0)"))
}
