% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexer.R
\name{.detokenise}
\alias{.detokenise}
\title{Reconstruct a formula string from a token list}
\usage{
.detokenise(tokens, prefix_eq = TRUE)
}
\arguments{
\item{tokens}{List of token objects.}

\item{prefix_eq}{Logical; prepend '=' if TRUE.}
}
\value{
Character scalar.
}
\description{
Reconstruct a formula string from a token list
}
\keyword{internal}
