% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{.get_sep}
\alias{.get_sep}
\title{Return the formula argument separator for a given locale.}
\usage{
.get_sep(locale)
}
\arguments{
\item{locale}{Two-letter locale code or NULL.}
}
\value{
";" or ","
}
\description{
Locales that use "," as their decimal separator use ";" in formulas.
}
\keyword{internal}
