% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{.strip_prefix}
\alias{.strip_prefix}
\title{Strip any OOXML prefix from a function name}
\usage{
.strip_prefix(fn)
}
\arguments{
\item{fn}{Character scalar.}
}
\value{
Clean function name without namespace prefix.
}
\description{
Strip any OOXML prefix from a function name
}
\keyword{internal}
