% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexer.R
\name{.tokenise}
\alias{.tokenise}
\title{Tokenise an Excel formula string}
\usage{
.tokenise(formula)
}
\arguments{
\item{formula}{Character scalar, optionally starting with '='.}
}
\value{
A list of token objects, each with fields \code{type} and \code{val}.
}
\description{
Tokenise an Excel formula string
}
\keyword{internal}
